add: product card component add: dashboard layout
